+++ src/util/figure.test.ts
@@ -22,9 +22,9 @@ describe('Unit: util/figure', () => {
         assert.deepStrictEqual(
           [...source.children].map(el => el.outerHTML),
           [
-            '<figure data-label="fig-a" data-group="fig" data-index="1" id="label:fig-1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
-            '<p><a rel="noopener" class="label" data-label="fig-a" href="#label:fig-1">Fig. 1</a></p>',
-            '<p><a rel="noopener" class="label" data-label="fig-a" href="#label:fig-1">Fig. 1</a></p>',
+            '<figure data-label="fig-a" data-group="fig" data-index="1" id="label:fig-a"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
+            '<p><a rel="noopener" class="label" data-label="fig-a" href="#label:fig-a">Fig. 1</a></p>',
+            '<p><a rel="noopener" class="label" data-label="fig-a" href="#label:fig-a">Fig. 1</a></p>',
           ]);
       }
     });
@@ -40,9 +40,9 @@ describe('Unit: util/figure', () => {
         assert.deepStrictEqual(
           [...source.children].map(el => el.outerHTML),
           [
-            '<figure data-label="fig-a" data-group="fig" data-index="1" id="label:fig-1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
-            '<figure data-label="table-a" data-group="table" data-index="1" id="label:table-1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Table. 1.</span><figcaption></figcaption></figure>',
-            '<figure data-label="fig-b" data-group="fig" data-index="2" id="label:fig-2"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 2.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-a" data-group="fig" data-index="1" id="label:fig-a"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="table-a" data-group="table" data-index="1" id="label:table-a"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Table. 1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-b" data-group="fig" data-index="2" id="label:fig-b"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 2.</span><figcaption></figcaption></figure>',
           ]);
       }
     });
@@ -60,11 +60,11 @@ describe('Unit: util/figure', () => {
         assert.deepStrictEqual(
           [...source.children].map(el => el.outerHTML),
           [
-            '<figure data-label="fig-a-0.0" data-group="fig" data-index="1.1" id="label:fig-1.1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.1.</span><figcaption></figcaption></figure>',
-            '<figure data-label="fig-b-0.0.0.0" data-group="fig" data-index="1.1.1.1" id="label:fig-1.1.1.1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.1.1.1.</span><figcaption></figcaption></figure>',
-            '<figure data-label="fig-c-0.0.0" data-group="fig" data-index="1.1.2" id="label:fig-1.1.2"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.1.2.</span><figcaption></figcaption></figure>',
-            '<figure data-label="fig-d" data-group="fig" data-index="1.1.3" id="label:fig-1.1.3"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.1.3.</span><figcaption></figcaption></figure>',
-            '<p><a rel="noopener" class="label" data-label="fig-b" href="#label:fig-1.1.1.1">Fig. 1.1.1.1</a></p>',
+            '<figure data-label="fig-a-0.0" data-group="fig" data-index="1.1" id="label:fig-a"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-b-0.0.0.0" data-group="fig" data-index="1.1.1.1" id="label:fig-b"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.1.1.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-c-0.0.0" data-group="fig" data-index="1.1.2" id="label:fig-c"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.1.2.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-d" data-group="fig" data-index="1.1.3" id="label:fig-d"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.1.3.</span><figcaption></figcaption></figure>',
+            '<p><a rel="noopener" class="label" data-label="fig-b" href="#label:fig-b">Fig. 1.1.1.1</a></p>',
           ]);
       }
     });
@@ -83,7 +83,7 @@ describe('Unit: util/figure', () => {
           [...source.children].map(el => el.outerHTML),
           [
             '<figure data-label="fig-1" data-group="fig" data-index="1" id="label:fig-1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
-            '<figure data-label="fig-a-0.0" data-group="fig" data-index="1.1" id="label:fig-1.1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-a-0.0" data-group="fig" data-index="1.1" id="label:fig-a"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.1.</span><figcaption></figcaption></figure>',
             '<figure data-label="$-3.1" data-group="$" data-index="3.1" id="label:$-3.1"><div class="figcontent"><div class="math notranslate">$$\n$$</div></div><span class="figindex">(3.1)</span><figcaption></figcaption></figure>',
             '<p><a rel="noopener" class="label" data-label="fig-2.0">$fig-2.0</a></p>',
             '<p><a rel="noopener" class="label" data-label="$-3.1" href="#label:$-3.1">(3.1)</a></p>',
@@ -121,22 +121,22 @@ describe('Unit: util/figure', () => {
             '<figure data-label="$-0.0" data-group="$" style="display: none;" data-index="0.0"><div class="figcontent"><div class="math notranslate">$$\n$$</div></div><span class="figindex">(0.0)</span><figcaption></figcaption></figure>',
             '<h2 id="index:0">0</h2>',
             '<figure data-label="fig-1" data-group="fig" data-index="1" id="label:fig-1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
-            '<figure data-label="fig-a-0.0" data-group="fig" data-index="1.1" id="label:fig-1.1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-a-0.0" data-group="fig" data-index="1.1" id="label:fig-a"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.1.</span><figcaption></figcaption></figure>',
             '<h2 id="index:0">0</h2>',
             '<blockquote><h2>0</h2></blockquote>',
-            '<figure data-label="fig-b" data-group="fig" data-index="2.1" id="label:fig-2.1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 2.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-b" data-group="fig" data-index="2.1" id="label:fig-b"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 2.1.</span><figcaption></figcaption></figure>',
             '<h2 id="index:0">0</h2>',
             '<figure data-label="$-0.0.0" data-group="$" style="display: none;" data-index="3.0.0"><div class="figcontent"><div class="math notranslate">$$\n$$</div></div><span class="figindex">(3.0.0)</span><figcaption></figcaption></figure>',
-            '<figure data-label="quote-a-0.0" data-group="quote" data-index="3.1" id="label:quote-3.1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Quote. 3.1.</span><figcaption></figcaption></figure>',
-            '<figure data-label="fig-c" data-group="fig" data-index="3.0.1" id="label:fig-3.0.1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 3.0.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="quote-a-0.0" data-group="quote" data-index="3.1" id="label:quote-a"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Quote. 3.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-c" data-group="fig" data-index="3.0.1" id="label:fig-c"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 3.0.1.</span><figcaption></figcaption></figure>',
             '<h2 id="index:0">0</h2>',
             '<figure data-label="$-0.1.0" data-group="$" style="display: none;" data-index="4.1.0"><div class="figcontent"><div class="math notranslate">$$\n$$</div></div><span class="figindex">(4.1.0)</span><figcaption></figcaption></figure>',
-            '<figure data-label="fig-d" data-group="fig" data-index="4.1.1" id="label:fig-4.1.1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 4.1.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-d" data-group="fig" data-index="4.1.1" id="label:fig-d"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 4.1.1.</span><figcaption></figcaption></figure>',
             '<h2 id="index:0">0</h2>',
             '<h2 id="index:0">0</h2>',
-            '<figure data-label="fig-e" data-group="fig" data-index="6.1" id="label:fig-6.1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 6.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-e" data-group="fig" data-index="6.1" id="label:fig-e"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 6.1.</span><figcaption></figcaption></figure>',
             '<figure data-label="$-5.0" data-group="$" style="display: none;" data-index="5.0"><div class="figcontent"><div class="math notranslate">$$\n$$</div></div><span class="figindex">(5.0)</span><figcaption></figcaption></figure>',
-            '<figure data-label="fig-d" data-group="fig" data-index="5.1" id="label:fig-5.1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 5.1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-d" data-group="fig" data-index="5.1" id="label:fig-d"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 5.1.</span><figcaption></figcaption></figure>',
           ]);
       }
     });
@@ -151,8 +151,8 @@ describe('Unit: util/figure', () => {
         assert.deepStrictEqual(
           [...source.children].map(el => el.outerHTML),
           [
-            '<figure data-label="$-a" data-group="$" data-index="1" id="label:$-1"><div class="figcontent"><div class="math notranslate">$$\n$$</div></div><span class="figindex">(1)</span><figcaption></figcaption></figure>',
-            '<p><a rel="noopener" class="label" data-label="$-a" href="#label:$-1">(1)</a></p>',
+            '<figure data-label="$-a" data-group="$" data-index="1" id="label:$-a"><div class="figcontent"><div class="math notranslate">$$\n$$</div></div><span class="figindex">(1)</span><figcaption></figcaption></figure>',
+            '<p><a rel="noopener" class="label" data-label="$-a" href="#label:$-a">(1)</a></p>',
           ]);
       }
     });
@@ -170,7 +170,7 @@ describe('Unit: util/figure', () => {
           [
             '<blockquote><blockquote><figure data-label="fig-1" data-group="fig" data-index="1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure></blockquote><figure data-label="fig-0" data-group="fig" data-index="0"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 0.</span><figcaption></figcaption></figure></blockquote>',
             '<aside class="example" data-type="markdown"><pre>~~~figure $fig-a\n&gt; \n\n~~~</pre><div><figure data-label="fig-a" data-group="fig" data-index="1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure></div><ol></ol><ol></ol></aside>',
-            '<figure data-label="fig-a" data-group="fig" data-index="1" id="label:fig-1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
+            '<figure data-label="fig-a" data-group="fig" data-index="1" id="label:fig-a"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
           ]);
       }
     });
@@ -186,9 +186,9 @@ describe('Unit: util/figure', () => {
         assert.deepStrictEqual(
           [...source.children].map(el => el.outerHTML),
           [
-            '<figure data-label="fig-a" data-group="fig" data-index="1" id="label:fig-1"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
-            '<p><a rel="noopener" class="label" data-label="fig-a" href="#label:fig-1">Fig. 1</a></p>',
-            '<p><a rel="noopener" class="label" data-label="fig-a" href="#label:fig-1">Fig. 1</a></p>',
+            '<figure data-label="fig-a" data-group="fig" data-index="1" id="label:fig-a"><div class="figcontent"><blockquote></blockquote></div><span class="figindex">Fig. 1.</span><figcaption></figcaption></figure>',
+            '<p><a rel="noopener" class="label" data-label="fig-a" href="#label:fig-a">Fig. 1</a></p>',
+            '<p><a rel="noopener" class="label" data-label="fig-a" href="#label:fig-a">Fig. 1</a></p>',
           ]);
       }
     });