+++ src/util/figure.ts
@@ -37,7 +37,7 @@ export function figure(source: DocumentFragment | HTMLElement | ShadowRoot): voi
     void define(figindex, group === '$' ? `(${idx})` : `${capitalize(group)}. ${idx}.`);
     if (fig.matches('blockquote *')) continue;
     if (idx.endsWith('.0')) continue;
-    void fig.setAttribute('id', `label:${label.split('-', 1)[0]}-${idx}`);
+    void fig.setAttribute('id', `label:${label.split(/-0(?![0-9])/, 1)[0]}`);
     const query = isGroup(label) ? label.split('-').slice(0, -1).join('-') : label;
     for (const ref of source.querySelectorAll(`a.label[data-label="${query.replace(/[$.]/g, '\\$&')}"]`)) {
       void define(ref, { href: `#${fig.id}` }, figindex.textContent!.replace(/[.]$/, ''));